requirements.txt
# 🐺 Cyberhound B2B - Production Requirements
# Real scraping - no simulation

# HTTP client for real web requests
aiohttp>=3.8.0
aiohttp-socks>=0.8.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data validation
pydantic>=2.0.0

# Async utilities
asyncio-mqtt>=0.16.0

# Telegram bot (for HOTL)
python-telegram-bot>=20.0

# Database (for lead storage)
aiosqlite>=0.19.0

# HTTP/2 support
hypercorn>=0.14.0
httpx[http2]>=0.24.0

# Monitoring
prometheus-client>=0.17.0

# ========== DEVELOPMENT / TESTING ==========

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
aioresponses>=0.7.4

# Linting and formatting
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Type stubs
mypy-extensions>=1.0.0
types-aiofiles>=23.0.0
